Correction de divers bug: - ajout des dateWish des reponses des API - suppression du paramètre idUser en entrée de l'API Get UserAccount (le traimenent se base sur le token) - delete Wish et Watch

+++ MMA-ApplicationLayer/src/main/java/com/bcefit/projet/service/moviedb/IEpisodeServiceImpl.java
@@ -27,7 +27,7 @@ public class IEpisodeServiceImpl implements IEpisodeService {
         Episode episode = new Episode();
         episode = iEpisodeRepository.findByIdEpisode(idEpisode);
         if (episode == null){
-            Tv tv = iTmdbApiService.synchronizeTvDetailFromApiFromApi(idTv.intValue());
+            Tv tv = iTmdbApiService.synchronizeTvDetailFromApiFromApi(idTv);
             episode = iEpisodeRepository.findByIdEpisode(idEpisode);
         }
         return episode;
