+++ MMA-ApplicationLayer/src/main/java/com/bcefit/projet/service/watch/WatchSeasonServiceImpl.java
@@ -32,7 +32,7 @@ public class WatchSeasonServiceImpl implements IWatchSeasonService {
     @Override
     public List<WatchEpisode> createWatchEpisodeBySeasonId(Long idTv, Long idSeason, UserAccount userAccount) {
         // Synchroniser la base des Tv/Season/Episode
-        iTmdbApiService.synchronizeTvDetailFromApiFromApi(idTv.intValue());
+        iTmdbApiService.synchronizeTvDetailFromApiFromApi(idTv);
 
         //création de la liste de tous les épisodes de la Season
         List<Episode> episodeListForTvAndSeason = iEpisodeService.getAllEpisodeByIdTvAndIdSeason(idTv, idSeason);
